% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeTimeline.R
\name{closeTimeline}
\alias{closeTimeline}
\title{Close Timeline Session}
\usage{
closeTimeline()
}
\value{
No devuelve ningún valor, pero imprime mensajes en la consola sobre el resultado
de la operación.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función cierra la sesión de la línea de tiempo de Twitter previamente abierta
con la función \code{openTimeline()}. Intenta cerrar la sesión del navegador y elimina
el objeto 'timeline' del entorno global.
}
\details{
La función realiza las siguientes acciones:
\enumerate{
\item Intenta cerrar la sesión del navegador asociada al objeto 'timeline'.
\item Espera un segundo para asegurar que la sesión se cierre correctamente.
\item Elimina el objeto 'timeline' del entorno global.
\item Muestra un mensaje de éxito si todas las operaciones se realizan correctamente.
}

Si ocurre algún error durante el proceso, se captura y se muestra un mensaje de error.
}
\note{
Esta función asume que existe un objeto 'timeline' en el entorno global,
creado previamente por la función \code{openTimeline()}. Si el objeto no existe,
se producirá un error.
}
\examples{
\dontrun{
# Primero, abrir una línea de tiempo
openTimeline("rstatstweet")

# Luego, cerrar la línea de tiempo
closeTimeline()
}

}
\seealso{
\code{\link{openTimeline}} para abrir una línea de tiempo de Twitter.
}
