% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthentication.R
\name{getAuthentication}
\alias{getAuthentication}
\title{Authenticate and Log In to Twitter}
\usage{
getAuthentication()
}
\value{
Esta función no devuelve ningún valor, solo autentica y realiza el inicio de sesión en Twitter.
}
\description{
Esta función automatiza el proceso de autenticación y inicio de sesión en Twitter utilizando las credenciales proporcionadas.
}
\details{
Esta función realiza la autenticación en Twitter y luego inicia sesión.


Utiliza la librería rvest para realizar la interacción con la página de inicio de sesión de Twitter.
}
\examples{
\dontrun{
getAuthentication()
}

}
\references{
Para obtener más información sobre la autenticación en Twitter, consulta la documentación del paquete.
}
