% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsData.R
\name{getTweetsData}
\alias{getTweetsData}
\title{Get Tweets Data}
\usage{
getTweetsData(
  urls_tweets,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS")
)
}
\arguments{
\item{urls_tweets}{Vector de URLs de tweets de los cuales se desea obtener datos.}

\item{xuser}{Nombre de usuario de Twitter para autenticaci칩n. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contrase침a de Twitter para autenticaci칩n. Por defecto es el valor de la variable de entorno del sistema PASS.}
}
\value{
Un tibble que contiene los datos de los tweets recuperados.
}
\description{
Esta funci칩n recupera datos de tweets a partir de URLs de tweets proporcionadas.
}
\examples{
\dontrun{
getTweetsData(urls_tweets = "https://twitter.com/estacion_erre/status/1788929978811232537")
}

}
