% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsFullSearch.R
\name{getTweetsFullSearch}
\alias{getTweetsFullSearch}
\title{Get Tweets from a Full Search}
\usage{
getTweetsFullSearch(
  search_all = "R Project",
  search_exact = NULL,
  search_any = NULL,
  no_search = NULL,
  hashtag = NULL,
  lan = NULL,
  from = NULL,
  to = NULL,
  men = NULL,
  rep = 0,
  fav = 0,
  rt = 0,
  timeout = 10,
  n_tweets = 100,
  since = Sys.Date() - 7,
  until = Sys.Date(),
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd(),
  save = TRUE
)
}
\arguments{
\item{search_all}{Cadena de texto. Busca tweets que contengan todas estas palabras (por defecto "R Project").}

\item{search_exact}{Cadena de texto. Busca tweets que contengan esta frase exacta (por defecto NULL).}

\item{search_any}{Cadena de texto. Busca tweets que contengan cualquiera de estas palabras (por defecto NULL).}

\item{no_search}{Cadena de texto. Excluye tweets que contengan estas palabras (por defecto NULL).}

\item{hashtag}{Cadena de texto. Busca tweets con estos hashtags (por defecto NULL).}

\item{lan}{Cadena de texto. Filtra tweets por idioma (por defecto NULL).}

\item{from}{Cadena de texto. Busca tweets de estos usuarios (por defecto NULL).}

\item{to}{Cadena de texto. Busca tweets dirigidos a estos usuarios (por defecto NULL).}

\item{men}{Cadena de texto. Busca tweets que mencionan a estos usuarios (por defecto NULL).}

\item{rep}{Número entero. Número mínimo de respuestas que debe tener un tweet (por defecto 0).}

\item{fav}{Número entero. Número mínimo de favoritos que debe tener un tweet (por defecto 0).}

\item{rt}{Número entero. Número mínimo de retweets que debe tener un tweet (por defecto 0).}

\item{timeout}{Número entero. Tiempo de espera en segundos entre solicitudes (por defecto 10).}

\item{n_tweets}{Número entero. Número máximo de tweets a recolectar (por defecto 100).}

\item{since}{Fecha. Fecha de inicio para la búsqueda (por defecto 7 días antes de la fecha actual).}

\item{until}{Fecha. Fecha de fin para la búsqueda (por defecto la fecha actual).}

\item{xuser}{Cadena de texto. Nombre de usuario para la autenticación en Twitter (por defecto se toma de la variable de entorno USER).}

\item{xpass}{Cadena de texto. Contraseña para la autenticación en Twitter (por defecto se toma de la variable de entorno PASS).}

\item{dir}{Cadena de texto. Directorio donde se guardarán los resultados (por defecto el directorio de trabajo actual).}

\item{save}{Lógico. Indica si se debe guardar el resultado en un archivo RDS (por defecto TRUE).}
}
\value{
Un tibble con los tweets recolectados, incluyendo las columnas:
\item{art_html}{HTML del artículo del tweet}
\item{fecha}{Fecha y hora del tweet}
\item{user}{Nombre de usuario del autor del tweet}
\item{tweet}{Texto del tweet}
\item{url}{URL del tweet}
\item{fecha_captura}{Fecha y hora de la captura del tweet}
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función realiza una búsqueda avanzada de tweets en Twitter (X) utilizando
varios criterios de búsqueda y recolecta los tweets que coinciden con estos criterios.
}
\details{
La función primero intenta autenticarse en Twitter utilizando las credenciales proporcionadas.
Luego, construye una URL de búsqueda basada en los parámetros proporcionados y realiza la búsqueda.
Los tweets se recolectan iterativamente, scrolleando la página hasta que se alcance el número
deseado de tweets o se agoten los intentos.
Los tweets recolectados se procesan para extraer la información relevante y, si save es TRUE,
se guardan en un archivo RDS.
}
\note{
Esta función requiere una conexión a Internet y credenciales válidas de Twitter.
}
\examples{
\dontrun{
tweets <- getTweetsFullSearch(
  search_all = "clima cambio",
  hashtag = "#medioambiente",
  lan = "es",
  n_tweets = 100,
  since = Sys.Date() - 30,
  save = TRUE
)

# Sin guardar los resultados
tweets <- getTweetsFullSearch(
  search_all = "clima cambio",
  hashtag = "#medioambiente",
  lan = "es",
  n_tweets = 100,
  since = Sys.Date() - 30,
  save = FALSE
)
}

}
