% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsHistoricalHashtag.R
\name{getTweetsHistoricalHashtag}
\alias{getTweetsHistoricalHashtag}
\title{Get Historical Tweets with a Specific Hashtag}
\usage{
getTweetsHistoricalHashtag(
  hashtag = "#rstats",
  timeout = 10,
  n_tweets = 100,
  since = "2018-10-26",
  until = "2018-10-30",
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd()
)
}
\arguments{
\item{hashtag}{Hashtag de Twitter del cual se desean recuperar los tweets históricos. Por defecto es "#rstats".}

\item{timeout}{Tiempo de espera entre solicitudes en segundos. Por defecto es 10.}

\item{n_tweets}{El número máximo de tweets a recuperar. Por defecto es 100.}

\item{since}{Fecha de inicio para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2018-10-26".}

\item{until}{Fecha de fin para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2018-10-30".}

\item{xuser}{Nombre de usuarix de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{dir}{Directorio para guardar el archivo RDS con los datos de los tweets recolectados. Por defecto es el directorio de trabajo actual.}
}
\value{
Un tibble que contiene los datos de tweets recuperados, junto con la fecha, usuario, contenido del tweet y URL del tweet.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera datos de los tweets que contienen un hashtag específico en Twitter,
basado en un rango de fechas especificado. Utiliza autenticación en Twitter mediante el
nombre de usuarix y la contraseña proporcionados, o los valores predeterminados de las
variables de entorno del sistema. Después de autenticar al usuarix, la función realiza
una búsqueda de tweets que contienen el hashtag especificado dentro del rango de fechas
definido por los parámetros \code{since} y \code{until}. Las URLs de los tweets encontrados se recogen
hasta alcanzar el número máximo de tweets especificado por el parámetro \code{n_tweets} o hasta que no
se encuentren nuevos tweets en varios intentos consecutivos. Los resultados se guardan en un
archivo con formato \code{.rds} en el directorio especificado por el parámetro \code{dir}.
}
\examples{
\dontrun{
getTweetsHistoricalHashtag(hashtag = "#rstats", n_tweets = 150, since = "2018-10-26", until = "2018-10-30")
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
