% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsHistoricalTimelineFor.R
\name{getTweetsHistoricalTimelineFor}
\alias{getTweetsHistoricalTimelineFor}
\title{Get Historical Tweets from User Timeline Iteratively}
\usage{
getTweetsHistoricalTimelineFor(
  iterations,
  username,
  n_tweets,
  since,
  until,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd(),
  system = "windows",
  sleep_time = 5 * 60
)
}
\arguments{
\item{iterations}{Número de iteraciones a realizar.}

\item{username}{Nombre de usuario de Twitter del cual se recolectarán los tweets.}

\item{n_tweets}{Número de tweets a recolectar por iteración.}

\item{since}{Fecha de inicio para la búsqueda (formato: "YYYY-MM-DD").}

\item{until}{Número de días a avanzar en cada iteración.}

\item{xuser}{Nombre de usuario de Twitter para autenticación (por defecto: variable de entorno del sistema "USER").}

\item{xpass}{Contraseña de Twitter para autenticación (por defecto: variable de entorno del sistema "PASS").}

\item{dir}{Directorio para guardar los tweets recolectados (por defecto: directorio de trabajo actual).}

\item{system}{Sistema operativo ("windows", "unix", o "mac").}

\item{sleep_time}{Tiempo de espera entre iteraciones en segundos (por defecto: 300 segundos).}
}
\value{
No devuelve un valor explícito, pero guarda los tweets recolectados en el directorio especificado.
}
\description{
Esta función realiza búsquedas históricas de tweets de la línea de tiempo de un usuario de forma iterativa,
permitiendo recolectar tweets en intervalos de tiempo específicos.
}
\details{
La función realiza las siguientes operaciones:
\enumerate{
\item Verifica e instala los paquetes necesarios.
\item Crea el directorio de destino si no existe.
\item Ejecuta búsquedas históricas de tweets de la línea de tiempo del usuario de forma iterativa.
\item Cierra el navegador después de cada iteración.
\item Espera un tiempo especificado entre iteraciones.
}
}
\examples{
\dontrun{
getTweetsHistoricalTimelineFor(
  iterations = 5,
  username = "rstatstweet",
  n_tweets = 10,
  since = "2018-07-01",
  until = 60,
  dir = "./datos/tweets",
  system = "windows",
  sleep_time = 10
)
}

}
