% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsReplies.R
\name{getTweetsReplies}
\alias{getTweetsReplies}
\title{Get Tweets Replies with Data}
\usage{
getTweetsReplies(
  url = "https://x.com/Picanumeros/status/1610715405705789442",
  n_tweets = 100,
  timeout = 2.5,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  view = FALSE,
  dir = getwd(),
  save = TRUE
)
}
\arguments{
\item{url}{URL del tweet del cual se quieren obtener las respuestas. Por defecto es "https://x.com/Picanumeros/status/1610715405705789442".}

\item{n_tweets}{El número máximo de tweets de respuestas a recuperar. Por defecto es 100.}

\item{timeout}{Tiempo de espera entre scrolls en segundos. Por defecto es 2.5.}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{view}{Ver el navegador. Por defecto es FALSE.}

\item{dir}{Directorio donde se guardará el archivo RDS con los datos recolectados. Por defecto es el directorio de trabajo actual.}

\item{save}{Lógico. Indica si se debe guardar el resultado en un archivo RDS (por defecto TRUE).}
}
\value{
Un data frame que contiene información sobre las respuestas al tweet especificado, incluyendo usuario, texto, fecha, URL y fecha de captura.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera las respuestas a un tweet específico en Twitter (ahora X),
incluyendo datos como el texto del tweet, usuario, fecha, y URL.
Utiliza web scraping para acceder a la página del tweet, iniciar sesión con las credenciales proporcionadas,
y recolectar la información de las respuestas al tweet.

El proceso incluye:
\enumerate{
\item Iniciar sesión en Twitter usando las credenciales proporcionadas.
\item Navegar a la URL del tweet especificado.
\item Extraer la información de las respuestas mediante scraping.
\item Continuar scrolling y recolectando datos hasta alcanzar el número deseado o no encontrar nuevas respuestas.
}

La función guarda los datos recolectados en un archivo RDS en el directorio especificado si el parámetro 'save' es TRUE,
y los devuelve como un data frame.
}
\note{
Esta función utiliza web scraping y puede ser sensible a cambios en la estructura de la página de Twitter.
}
\examples{
\dontrun{
getTweetsReplies(url = "https://x.com/Picanumeros/status/1610715405705789442", n_tweets = 130)

# Sin guardar los resultados
getTweetsReplies(url = "https://x.com/Picanumeros/status/1610715405705789442", n_tweets = 130, save = FALSE)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScraperR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
