% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsRetweets.R
\name{getTweetsRetweets}
\alias{getTweetsRetweets}
\title{Get Users Retweets with Data}
\usage{
getTweetsRetweets(
  url = "https://x.com/tipsder/status/1672311054922293254",
  n_users = 100,
  timeout = 2.5,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  view = FALSE,
  dir = getwd(),
  save = TRUE,
  open = FALSE
)
}
\arguments{
\item{url}{URL del tweet del cual se quieren obtener los retweets. Por defecto es "https://x.com/tipsder/status/1672311054922293254".}

\item{n_users}{El número máximo de users a recuperar. Por defecto es 100.}

\item{timeout}{Tiempo de espera entre scrolls en segundos. Por defecto es 2.5.}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{view}{Ver el navegador. Por defecto es FALSE.}

\item{dir}{Directorio donde se guardará el archivo RDS con los datos recolectados. Por defecto es el directorio de trabajo actual.}

\item{save}{Lógico. Indica si se debe guardar el resultado en un archivo RDS (por defecto TRUE).}

\item{open}{Lógico. Indica si se debe abrir una nueva sesión de login en Twitter (por defecto FALSE).}
}
\value{
Un data frame que contiene información sobre los users que rt el tweet especificado, incluyendo usuario, URL y fecha de captura.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera los retweets a un tweet específico en Twitter (ahora X),
incluyendo datos como el user name, name y URL.
Utiliza web scraping para acceder a la página del tweet, iniciar sesión con las credenciales proporcionadas,
y recolectar la información de los retweets al tweet.

El proceso incluye:
\enumerate{
\item Iniciar sesión en Twitter usando las credenciales proporcionadas (si open=TRUE).
\item Navegar a la URL del tweet especificado con "/retweets" para ver los retweets.
\item Extraer la información de los retweets mediante scraping.
\item Continuar scrolling y recolectando datos hasta alcanzar el número deseado o no encontrar nuevas citas.
}

La función guarda los datos recolectados en un archivo RDS en el directorio especificado si el parámetro 'save' es TRUE,
y los devuelve como un data frame.
}
\note{
Esta función utiliza web scraping y puede ser sensible a cambios en la estructura de la página de Twitter.
}
\examples{
\dontrun{
getTweetsRetweets(url = "https://x.com/tipsder/status/1672311054922293254", n_users = 20)

# Sin guardar los resultados
getTweetsRetweets(url = "https://x.com/tipsder/status/1672311054922293254", n_users = 20, save = FALSE)

# Sin abrir una nueva sesión de login
getTweetsRetweets(url = "https://x.com/tipsder/status/1672311054922293254", n_users = 20, open = TRUE)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScraperR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
