% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsSearchStreaming.R
\name{getTweetsSearchStreaming}
\alias{getTweetsSearchStreaming}
\title{Get Live Tweet by Search}
\usage{
getTweetsSearchStreaming(
  search = "#RStats",
  timeout = 10,
  n_tweets = 100,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd(),
  save = TRUE
)
}
\arguments{
\item{search}{La consulta de búsqueda para recuperar tweets. Por defecto es "#RStats".}

\item{timeout}{Tiempo de espera.}

\item{n_tweets}{El número máximo de tweets a recuperar. Por defecto es 100.}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{dir}{Directorio para guardar el archivo RDS con las URLs recolectadas.}

\item{save}{Lógico. Indica si se debe guardar el resultado en un archivo RDS (por defecto TRUE).}
}
\value{
Un vector que contiene las URLs de tweets recuperadas.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera tweets basados en una consulta de búsqueda en tiempo real en Twitter.
Utiliza autenticación en Twitter mediante el nombre de usuario y la contraseña proporcionados,
o los valores predeterminados de las variables de entorno del sistema. Después de autenticar al usuario,
la función realiza la búsqueda especificada por el parámetro \code{search} y recoge las URLs de los tweets que
coinciden con la consulta.
El proceso de recolección de URLs se ejecuta en un bucle que continúa hasta que se alcanza el número máximo
de URLs especificado por el parámetro \code{n_tweets} o hasta que se realizan varios intentos consecutivos sin
encontrar nuevas URLs, indicando que no hay más resultados disponibles en ese momento. La función incorpora
mecanismos de manejo de errores y tiempos de espera para asegurar que las conexiones y búsquedas se realicen
de manera robusta y continua.
Las URLs de los tweets recolectados se almacenan en un vector y se guardan en un archivo con formato \code{.rds}
en el directorio especificado por el parámetro \code{dir} si el parámetro \code{save} es TRUE. Este archivo se nombra
de manera única utilizando la consulta de búsqueda y la marca de tiempo del momento en que se realiza la
recolección, asegurando que no se sobrescriban archivos anteriores.
}
\examples{
\dontrun{
getTweetsSearchStreaming(search = "#RStats", n_tweets = 200)

# Sin guardar los resultados
getTweetsSearchStreaming(search = "#RStats", n_tweets = 200, save = FALSE)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
