% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsSearchStreamingFor.R
\name{getTweetsSearchStreamingFor}
\alias{getTweetsSearchStreamingFor}
\title{Get Tweets Iteratively}
\usage{
getTweetsSearchStreamingFor(
  iterations,
  search,
  n_tweets,
  dir,
  system = "unix",
  sleep_time = 300
)
}
\arguments{
\item{iterations}{Número de iteraciones a realizar}

\item{search}{Término de búsqueda para los tweets}

\item{n_tweets}{Número de tweets a recolectar en cada iteración}

\item{dir}{Directorio donde se guardarán los tweets}

\item{system}{Sistema operativo ('windows', 'unix', 'macOS')}

\item{sleep_time}{Tiempo de espera entre iteraciones en segundos}
}
\value{
No devuelve un valor, pero guarda los tweets en el directorio especificado
}
\description{
Esta función recolecta tweets de forma iterativa utilizando TweetScraperR,
con la opción de cerrar el navegador entre iteraciones.
}
\examples{
\dontrun{
getTweetsSearchStreamingFor(
  iterations = 5,
  search = "Milei",
  n_tweets = 10,
  dir = "./data/tweets",
  system = "unix",
  sleep_time = 5
)
}
}
