% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsTimeline.R
\name{getTweetsTimeline}
\alias{getTweetsTimeline}
\title{Get Tweets from User Timeline}
\usage{
getTweetsTimeline(
  username = "rstatstweet",
  n_tweets = 100,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS")
)
}
\arguments{
\item{username}{El nombre de usuario de Twitter del cual quieres obtener el timeline.}

\item{n_tweets}{El número máximo de tweets a obtener. Por defecto es 100.}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}
}
\value{
Un vector que contiene los tweets obtenidos.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera tweets del timeline de un usuario especificado en Twitter.
}
\examples{
\dontrun{
getTweetsTimeline(username = "rstatstweet", n_tweets = 200)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
