% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsUrlsSearch.R
\name{getTweetsUrlsSearch}
\alias{getTweetsUrlsSearch}
\title{Get Tweets URLs by Search}
\usage{
getTweetsUrlsSearch(
  search = "#RStats",
  n_urls = 100,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  max_retries = 3
)
}
\arguments{
\item{search}{La consulta de búsqueda para usar en la recuperación de tweets. Por defecto es "#RStats".}

\item{n_urls}{El número máximo de URLs de tweets a recuperar. Por defecto es 100.}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{max_retries}{número máximo de intentos de conexión.}
}
\value{
Un vector que contiene las URLs de tweets recuperadas.
}
\description{
Obtener URLs de Tweets por Búsqueda
}
\details{
Esta función recupera URLs de tweets basados en una consulta de búsqueda especificada en Twitter.
}
\examples{
\dontrun{
getTweetsUrlsSearch(search = "#RStats", n_urls = 200)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
