% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsUrlsTimeline.R
\name{getTweetsUrlsTimeline}
\alias{getTweetsUrlsTimeline}
\title{Get URLs of User Timeline Tweets}
\usage{
getTweetsUrlsTimeline(
  username = "rstatstweet",
  n_urls = 100,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS")
)
}
\arguments{
\item{username}{El nombre de usuario de Twitter del cual quieres obtener el timeline. Por defecto es "rstatstweet".}

\item{n_urls}{El número máximo de URLs de tweets a obtener. Por defecto es 100.}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}
}
\value{
Un vector que contiene las URLs de tweets obtenidas.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera URLs de tweets del timeline de un usuario especificado en Twitter.
}
\examples{
\dontrun{
getTweetsUrlsTimeline(username = "rstatstweet", n_urls = 200)
}

}
