% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlsHistoricalHashtag.R
\name{getUrlsHistoricalHashtag}
\alias{getUrlsHistoricalHashtag}
\title{Get Historical URLs of Tweets with a Specific Hashtag}
\usage{
getUrlsHistoricalHashtag(
  hashtag = "#rstats",
  timeout = 10,
  n_urls = 100,
  since = "2018-10-26",
  until = "2018-10-30",
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd()
)
}
\arguments{
\item{hashtag}{Hashtag de Twitter del cual se desean recuperar los tweets históricos. Por defecto es "#rstats".}

\item{timeout}{Tiempo de espera entre solicitudes en segundos. Por defecto es 10.}

\item{n_urls}{El número máximo de URLs de tweets a recuperar. Por defecto es 100.}

\item{since}{Fecha de inicio para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2018-10-26".}

\item{until}{Fecha de fin para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2018-10-30".}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{dir}{Directorio para guardar el archivo RDS con las URLs recolectadas. Por defecto es el directorio de trabajo actual.}
}
\value{
Un tibble que contiene las URLs de tweets recuperadas, junto con la fecha, usuario, contenido del tweet y URL del tweet.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera URLs de tweets históricos que contienen un hashtag específico en Twitter.
}
\examples{
\dontrun{
getUrlsHistoricalHashtag(hashtag = "#rstats", n_urls = 150, since = "2018-10-26", until = "2018-10-30")
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
