% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlsHistoricalSearch.R
\name{getUrlsHistoricalSearch}
\alias{getUrlsHistoricalSearch}
\title{Get Historical URLs of Tweets from a Specific Search}
\usage{
getUrlsHistoricalSearch(
  search = "R Project",
  timeout = 10,
  n_urls = 100,
  since = "2018-10-26",
  until = "2023-10-30",
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd()
)
}
\arguments{
\item{search}{Término de búsqueda para los tweets deseados. Por defecto es "R Project".}

\item{timeout}{Tiempo de espera entre solicitudes en segundos. Por defecto es 10.}

\item{n_urls}{El número máximo de URLs de tweets a recuperar. Por defecto es 100.}

\item{since}{Fecha de inicio para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2018-10-26".}

\item{until}{Fecha de fin para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2023-10-30".}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{dir}{Directorio para guardar el archivo RDS con las URLs recolectadas. Por defecto es el directorio de trabajo actual.}
}
\value{
Un tibble que contiene las URLs de tweets recuperadas, junto con la fecha, usuario, contenido del tweet y URL del tweet.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera URLs de tweets históricos que corresponden a una búsqueda específica en Twitter.
}
\examples{
\dontrun{
getUrlsHistoricalSearch(search = "R Project", n_urls = 50, since = "2018-10-26", until = "2023-10-30")
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
