% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsersData.R
\name{getUsersData}
\alias{getUsersData}
\title{Get Users Data}
\usage{
getUsersData(
  urls_users,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS")
)
}
\arguments{
\item{urls_users}{Vector de URLs de users de los cuales se desea obtener datos.}

\item{xuser}{Nombre de usuario de Twitter para autenticaci칩n. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contrase침a de Twitter para autenticaci칩n. Por defecto es el valor de la variable de entorno del sistema PASS.}
}
\value{
Un tibble que contiene los datos de los users recuperados.
}
\description{
Esta funci칩n recupera datos de users a partir de URLs de users proporcionadas.
}
\examples{
\dontrun{
getUsersData(urls_users = "https://x.com/estacion_erre")
}

}
