% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openTimeline.R
\name{openTimeline}
\alias{openTimeline}
\title{Open Timeline User}
\usage{
openTimeline(username = "rstatstweet", view = TRUE)
}
\arguments{
\item{username}{Character. El nombre de usuario de Twitter cuya línea de tiempo se desea abrir.
Por defecto es "rstatstweet".}

\item{view}{Logical. Si es TRUE (por defecto), muestra la vista de la página web.
Si es FALSE, solo crea el objeto sin mostrar la vista.}
}
\value{
Si view es TRUE, devuelve la vista de la página web.
Si view es FALSE, devuelve un mensaje indicando que se ha creado el objeto "timeline".
En caso de error después de tres intentos, devuelve NULL.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función abre la línea de tiempo de un usuario específico de Twitter (X) y crea un objeto
que contiene la información de la página. Intenta abrir la página hasta tres veces en caso de error.
}
\details{
La función utiliza \code{rvest::read_html_live()} para leer la página web de la línea de tiempo
del usuario especificado. Crea un objeto global llamado "timeline" que contiene la información
de la página. Si ocurre un error al intentar abrir la página, la función reintentará hasta
tres veces antes de fallar.
}
\note{
Esta función requiere una conexión a Internet activa y puede estar sujeta a las limitaciones
de acceso impuestas por Twitter (X).
}
\examples{
\dontrun{
# Abrir la línea de tiempo de un usuario específico y mostrar la vista
openTimeline("hadleywickham")

# Crear el objeto timeline sin mostrar la vista
openTimeline("rstudio", view = FALSE)
}

}
