% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openTwitter.R
\name{openTwitter}
\alias{openTwitter}
\title{Open Twitter Login Page}
\usage{
openTwitter(view = TRUE)
}
\arguments{
\item{view}{Logical. If TRUE (default), returns a view of the HTML instance. If FALSE, returns the HTML instance itself.}
}
\value{
Si \code{view} es TRUE, devuelve una vista de la instancia HTML de la página de inicio de sesión de Twitter.
Si \code{view} es FALSE, devuelve la instancia HTML directamente.
En ambos casos, la información de la sesión se guarda en la variable global \code{twitter} para su uso posterior.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función permite abrir la página de inicio de sesión de Twitter en un navegador web y
guardar la instancia HTML en el entorno global. Es útil para iniciar el proceso de autenticación
antes de realizar la recolección de datos de Twitter.
}
\examples{
\dontrun{
# Para obtener la vista (comportamiento predeterminado)
openTwitter()

# Para obtener el objeto twitter sin la vista
twitter <- openTwitter(view = FALSE)
}

}
