% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmojis.R
\name{plotEmojis}
\alias{plotEmojis}
\title{Create Bar Chart of Emoticons in Tweets}
\usage{
plotEmojis(df, top_n = 10, fill = "skyblue", color = "grey30")
}
\arguments{
\item{df}{Un dataframe que contiene una columna 'emoticones' con listas de emoticones.}

\item{top_n}{Número de emoticones más frecuentes a mostrar (por defecto 10).}

\item{fill}{Color de las barras en el gráfico (por defecto "skyblue").}

\item{color}{Color de los bordes de las barras en el gráfico (por defecto "grey30").}
}
\value{
Un objeto ggplot con el gráfico de barras.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función toma un dataframe de tweets y crea un gráfico de barras
mostrando la frecuencia de los emoticones utilizados en los tweets.
}
\examples{
\dontrun{
df <- data.frame(emoticones = I(list(
  c("\U0001F60A", "\U0001F602", "\U0001F602"),
  c("\U0001F60A"),
  c("\U0001F602", "\U0001F60D"),
  character(0)
)))
plotEmojis(df)
}

}
