% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTime.R
\name{plotTime}
\alias{plotTime}
\title{Create Line Graph of Tweets by Time}
\usage{
plotTime(df, group_by = "hour", color = "blue")
}
\arguments{
\item{df}{Un dataframe que contiene una columna 'fecha' con las fechas de los tweets.}

\item{group_by}{Una cadena que indica cómo agrupar las fechas.
Opciones válidas son "hour", "day", "week", "month", "year".}

\item{color}{Color de la línea en el gráfico (por defecto "blue").}
}
\value{
Un objeto ggplot con el gráfico de líneas.
}
\description{
Esta función toma un dataframe de tweets y crea un gráfico de líneas
mostrando la frecuencia de tweets a lo largo del tiempo, con opciones
para agrupar por hora, día, semana, mes o año.
}
\examples{

df <- data.frame(fecha = c(
  seq(as.POSIXct("2023-01-02 00:00:00"), by = "hour", length.out = 3),
  seq(as.POSIXct("2023-01-02 01:00:00"), by = "hour", length.out = 2),
  seq(as.POSIXct("2023-01-02 03:00:00"), by = "hour", length.out = 3),
  seq(as.POSIXct("2023-01-02 01:00:00"), by = "hour", length.out = 2),
  seq(as.POSIXct("2023-01-02 03:00:00"), by = "hour", length.out = 3),
  seq(as.POSIXct("2023-01-02 01:00:00"), by = "hour", length.out = 2),
  seq(as.POSIXct("2023-01-02 03:00:00"), by = "hour", length.out = 3)
))
plotTime(df, group_by = "hour")

}
