% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWords.R
\name{plotWords}
\alias{plotWords}
\title{Create Word Cloud from Tweets}
\usage{
plotWords(
  df,
  min_freq = 3,
  max_words = 100,
  random_order = FALSE,
  colors = "random-dark",
  size = 0.3,
  lang = "es",
  sw = NULL
)
}
\arguments{
\item{df}{Un dataframe que contiene una columna 'texto' o 'tweet' con el contenido de los tweets.}

\item{min_freq}{Frecuencia mínima de palabras para incluir en la nube (por defecto 3).}

\item{max_words}{Número máximo de palabras a incluir en la nube (por defecto 100).}

\item{random_order}{Booleano, si las palabras deben ordenarse aleatoriamente (por defecto FALSE).}

\item{colors}{Vector de colores para las palabras (por defecto 'random-dark').}

\item{size}{Tamaño de la fuente (por defecto 0.3).}

\item{lang}{idioma para las stopwords 'es', 'en', 'de', 'pt', etc. (por defecto 'es').}

\item{sw}{vector de palabras extras para sumar a la lista de stopwords (por defecto NULL).}
}
\value{
Un objeto de tipo wordcloud2.
}
\description{
Esta función toma un dataframe de tweets y crea una nube de palabras
basada en el contenido de la columna 'texto' o 'tweet'.
}
\examples{

df <- data.frame(texto = c("Este es un tweet de ejemplo", "Otro tweet para la nube de palabras"))
plotWords(df, min_freq = 1)
}
