% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsData.R
\name{getTweetsData}
\alias{getTweetsData}
\title{Get Tweets Data}
\usage{
getTweetsData(
  urls_tweets,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd(),
  save = TRUE
)
}
\arguments{
\item{urls_tweets}{Vector de URLs de tweets de los cuales se desea obtener datos.}

\item{xuser}{Nombre de usuarix de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{dir}{directorio para guardar el RDS con las URLs recolectadas}

\item{save}{Logical. Indica si se debe guardar el resultado en un archivo RDS. Por defecto es TRUE.}
}
\value{
Un tibble que contiene los datos de los tweets recuperados.

\itemize{
\item \code{tweets_recuperados}: Un tibble con los datos de los tweets recuperados, incluyendo la fecha, nombre de usuario, texto, respuestas, reposts, me gusta, URLs asociadas y otras informaciones recopiladas.
\item \code{tweets_borrados}: Un vector con las URLs de los tweets que fueron detectados como borrados.
\item \code{tweets_a_reprocesar}: Un vector con las URLs de los tweets que no pudieron ser procesados exitosamente y necesitan ser reprocesados.
\item \code{errores}: Un vector con los mensajes de error recopilados durante el proceso de recolección de datos.
}
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función permite recuperar y procesar datos de tweets a partir de un vector de URLs
de tweets proporcionadas. Utilizando las credenciales de unx usuarix de Twitter, la función
realiza la autenticación en Twitter y extrae información detallada de cada tweet. Los datos
extraídos incluyen la fecha del tweet, el nombre de usuarix que lo publicó, el texto del tweet,
las respuestas, reposts, me gusta, URLs asociadas, y otra información relevante.
La función también maneja tweets borrados y errores durante el proceso de recolección, y
clasifica las URLs de los tweets en tres categorías: tweets recuperados, tweets borrados, y
tweets que necesitan ser reprocesados. Si el parámetro 'save' es TRUE, los datos recopilados
se guardan en un archivo RDS en el directorio especificado por le usuarix.
}
\examples{
\dontrun{
getTweetsData(urls_tweets = "https://twitter.com/estacion_erre/status/1788929978811232537")
getTweetsData(urls_tweets = "https://twitter.com/estacion_erre/status/1788929978811232537", save = FALSE)
}

}
