% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsHashtags.R
\name{getTweetsHashtags}
\alias{getTweetsHashtags}
\title{Get Hashtags from Tweets}
\usage{
getTweetsHashtags(df)
}
\arguments{
\item{df}{Un dataframe que contiene una columna 'texto' con el contenido de los tweets.}
}
\value{
Un dataframe con una nueva columna 'hashtags' que contiene una lista de hashtags para cada tweet.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función toma un dataframe de tweets y extrae todos los hashtags
del campo 'texto', añadiéndolos como una nueva columna al dataframe.
}
\examples{
df <- data.frame(texto = c("Este es un #tweet con #hashtags", "Este no tiene hashtags", "Otro #ejemplo"))
getTweetsHashtags(df)

}
