% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsHistoricalSearchFor.R
\name{getTweetsHistoricalSearchFor}
\alias{getTweetsHistoricalSearchFor}
\title{Get Historical Tweets Iteratively}
\usage{
getTweetsHistoricalSearchFor(
  iterations,
  search,
  n_tweets,
  since,
  until,
  interval_unit = "days",
  live = TRUE,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd(),
  system = "windows",
  sleep_time = 5 * 60
)
}
\arguments{
\item{iterations}{Número de iteraciones a realizar.}

\item{search}{Término de búsqueda para los tweets.}

\item{n_tweets}{Número de tweets a recolectar por iteración.}

\item{since}{Fecha y hora de inicio para la búsqueda (formato: "YYYY-MM-DD_HH:MM:SS_UTC").}

\item{until}{Número de unidades de tiempo a avanzar en cada iteración.}

\item{interval_unit}{Unidad de tiempo para el intervalo ("days", "hours", o "minutes").}

\item{live}{Booleano que indica si se debe realizar una búsqueda en vivo (por defecto: TRUE).}

\item{xuser}{Nombre de usuario de Twitter para autenticación (por defecto: variable de entorno del sistema "USER").}

\item{xpass}{Contraseña de Twitter para autenticación (por defecto: variable de entorno del sistema "PASS").}

\item{dir}{Directorio para guardar los tweets recolectados (por defecto: directorio de trabajo actual).}

\item{system}{Sistema operativo ("windows", "unix", o "mac").}

\item{sleep_time}{Tiempo de espera entre iteraciones en segundos (por defecto: 300 segundos).}
}
\value{
No devuelve un valor explícito, pero guarda los tweets recolectados en el directorio especificado.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función realiza búsquedas históricas de tweets de forma iterativa,
permitiendo recolectar tweets en intervalos de tiempo específicos.
}
\details{
La función realiza las siguientes operaciones:
\enumerate{
\item Valida el formato de la fecha y hora de inicio.
\item Crea el directorio de destino si no existe.
\item Ejecuta búsquedas históricas de tweets de forma iterativa.
\item Calcula la fecha y hora de finalización para cada iteración basándose en el intervalo especificado.
\item Cierra el navegador después de cada iteración según el sistema operativo especificado.
\item Espera un tiempo especificado entre iteraciones.
}
}
\examples{
\dontrun{
getTweetsHistoricalSearchFor(
  iterations = 5,
  search = "cambio climático",
  n_tweets = 1000,
  since = "2023-01-01_15:55:00_UTC",
  until = 7,
  interval_unit = "days",
  live = TRUE,
  dir = "./datos/tweets",
  system = "windows",
  sleep_time = 300
)
}

}
