% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsSearchStreamingFor.R
\name{getTweetsSearchStreamingFor}
\alias{getTweetsSearchStreamingFor}
\title{Get Iterative Tweets in Streaming}
\usage{
getTweetsSearchStreamingFor(
  iterations,
  search,
  n_tweets,
  dir,
  system = "unix",
  kill_system = FALSE,
  sleep_time = 300
)
}
\arguments{
\item{iterations}{Número de iteraciones a realizar}

\item{search}{Término de búsqueda para los tweets}

\item{n_tweets}{Número de tweets a recolectar en cada iteración}

\item{dir}{Directorio donde se guardarán los tweets}

\item{system}{Sistema operativo ('windows', 'unix', 'macOS')}

\item{kill_system}{Booleano que indica si se debe cerrar el navegador después de cada iteración (por defecto: FALSE)}

\item{sleep_time}{Tiempo de espera entre iteraciones en segundos}
}
\value{
No devuelve un valor, pero guarda los tweets en el directorio especificado
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recolecta tweets de forma iterativa utilizando TweetScraperR,
con la opción de cerrar el navegador entre iteraciones.
}
\examples{
\dontrun{
getTweetsSearchStreamingFor(
  iterations = 5,
  search = "Milei",
  n_tweets = 10,
  dir = "./data/tweets",
  system = "unix",
  kill_system = FALSE,
  sleep_time = 5
)
}
}
