% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsSentiments.R
\name{getTweetsSentiments}
\alias{getTweetsSentiments}
\title{Analyze sentiments of tweets}
\usage{
getTweetsSentiments(
  tweets,
  api_key = Sys.getenv("OPENAI_API_KEY"),
  model = "gpt-4o-mini",
  dir = getwd()
)
}
\arguments{
\item{tweets}{Un vector de caracteres que contiene los tweets a analizar.}

\item{api_key}{Una cadena de caracteres con la clave de API de OpenAI. Por defecto,
se intenta obtener de la variable de entorno OPENAI_API_KEY.}

\item{model}{Una cadena de caracteres que especifica el modelo de OpenAI a utilizar.
Por defecto es "gpt-4o-mini".}

\item{dir}{Una cadena de caracteres que especifica el directorio donde se guardar치
el archivo RDS con los resultados. Por defecto es el directorio de trabajo actual.}
}
\value{
Un tibble con los resultados del an치lisis para cada tweet. Cada fila
contiene el tweet original y los resultados del an치lisis (tono, sentimiento,
presencia de expresiones de odio, si est치 direccionado, si contiene un llamado
a la acci칩n y una explicaci칩n detallada).
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta funci칩n toma un vector de tweets y utiliza la API de OpenAI para realizar
un an치lisis de sentimiento detallado de cada tweet. El an치lisis incluye el tono,
el sentimiento general, la presencia de expresiones de odio, si el tweet est치
direccionado a alguien espec칤fico y si contiene un llamado a la acci칩n.
}
\details{
La funci칩n realiza las siguientes operaciones:
\enumerate{
\item Verifica que se haya proporcionado una clave de API v치lida.
\item Define un prompt detallado para el an치lisis de los tweets.
\item Define una funci칩n interna \code{analyze_tweet} que procesa cada tweet individualmente.
\item Utiliza \code{purrr::map_dfr} para aplicar \code{analyze_tweet} a cada tweet en el vector de entrada.
\item Guarda los resultados en un archivo RDS en el directorio especificado.
\item Devuelve los resultados como un tibble.
}

La funci칩n utiliza la API de OpenAI para realizar el an치lisis de sentimiento,
por lo que requiere una conexi칩n a Internet y una clave de API v치lida.
}
\examples{
\dontrun{
tweets <- c("춰Qu칠 d칤a tan maravilloso! 游땕", "Odio este producto, nunca lo compren. 游")
resultados <- getTweetsSentiments(tweets)
print(resultados)
}

}
