% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetsTimeline.R
\name{getTweetsTimeline}
\alias{getTweetsTimeline}
\title{Get Tweets from User Timeline}
\usage{
getTweetsTimeline(
  username = "rstatstweet",
  n_tweets = 100,
  view = FALSE,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  mailx = "tweetscraperr@gmail.com",
  dir = getwd(),
  save = TRUE
)
}
\arguments{
\item{username}{El nombre de usuarix de Twitter del cual quieres obtener el timeline.}

\item{n_tweets}{El número máximo de tweets a obtener. Por defecto es 100.}

\item{view}{Mostrar una vista en vivo de Twitter/X TRUE o FALSE}

\item{xuser}{Nombre de usuarix de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{mailx}{Dirección de e-mail para la autenticación. Tiene que ser la misma que la usada en Twitter/X.}

\item{dir}{El directorio donde se guardará el archivo de salida. Por defecto es el directorio de trabajo actual.}

\item{save}{Lógico. Indica si se debe guardar el resultado en un archivo RDS (por defecto TRUE).}
}
\value{
Un tibble que contiene los tweets obtenidos.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera tweets del timeline de unx usuarix especificadx en Twitter.
La función inicia sesión en Twitter utilizando las credenciales proporcionadas,
navega al perfil de le usuarix especificadx, y recopila hasta \code{n_tweets} tweets.
El proceso de recolección se detiene si se alcanza el número máximo de tweets
especificado o después de alcanzar los 600 tweets con el desplazamiento (scroll).
}
\examples{
\dontrun{
getTweetsTimeline(username = "rstatstweet", n_tweets = 200)

# Sin guardar los resultados
getTweetsTimeline(username = "rstatstweet", n_tweets = 200, save = FALSE)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
