% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlsHistoricalTimeline.R
\name{getUrlsHistoricalTimeline}
\alias{getUrlsHistoricalTimeline}
\title{Get Historical Tweet URLs from a User Timeline}
\usage{
getUrlsHistoricalTimeline(
  username = "rstatstweet",
  timeout = 10,
  n_urls = 100,
  since = "2018-10-26",
  until = "2018-10-30",
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd(),
  save = TRUE
)
}
\arguments{
\item{username}{Nombre de usuarix de Twitter del que se desean recuperar los tweets. Por defecto es "rstatstweet".}

\item{timeout}{Tiempo de espera entre solicitudes en segundos. Por defecto es 10.}

\item{n_urls}{El número máximo de URLs de tweets a recuperar. Por defecto es 100.}

\item{since}{Fecha de inicio para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2018-10-26".}

\item{until}{Fecha de fin para la búsqueda de tweets (en formato "YYYY-MM-DD"). Por defecto es "2018-10-30".}

\item{xuser}{Nombre de usuarix de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{dir}{Directorio para guardar el archivo RDS con las URLs recolectadas. Por defecto es el directorio de trabajo actual.}

\item{save}{Lógico. Indica si se debe guardar el resultado en un archivo RDS (por defecto TRUE).}
}
\value{
Un vector que contiene las URLs de tweets recuperadas.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función recupera URLs de tweets del timeline histórico de unx usuarix en Twitter,
basado en un rango de fechas especificado. Utiliza autenticación en Twitter mediante
el nombre de usuarix y la contraseña proporcionados, o los valores predeterminados de
las variables de entorno del sistema. Después de autenticar al usuarix, la función
realiza una búsqueda de tweets publicados por le usuarix especificadx dentro del rango
de fechas definido por los parámetros \code{since} y \code{until}. Las URLs de los tweets encontrados
se recogen hasta alcanzar el número máximo de URLs especificado por el parámetro \code{n_urls} o
hasta que no se encuentren nuevas URLs en varios intentos consecutivos. Los resultados se
guardan en un archivo con formato \code{.rds} en el directorio especificado por el parámetro \code{dir}
si el parámetro \code{save} es TRUE.
}
\examples{
\dontrun{
getUrlsHistoricalTimeline(username = "rstatstweet", n_urls = 50, since = "2018-10-26", until = "2018-10-30")

# Sin guardar los resultados
getUrlsHistoricalTimeline(username = "rstatstweet", n_urls = 50, since = "2018-10-26", until = "2018-10-30", save = FALSE)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
