% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlsSearchStreaming.R
\name{getUrlsSearchStreaming}
\alias{getUrlsSearchStreaming}
\title{Get Live Tweet URLs by Search}
\usage{
getUrlsSearchStreaming(
  search = "#RStats",
  timeout = 10,
  n_urls = 100,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS")
)
}
\arguments{
\item{search}{La consulta de búsqueda para recuperar tweets. Por defecto es "#RStats".}

\item{n_urls}{El número máximo de URLs de tweets a recuperar. Por defecto es 100.}

\item{xuser}{Nombre de usuario de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}
}
\value{
Un vector que contiene las URLs de tweets recuperadas.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Esta función recupera URLs de tweets basados en una consulta de búsqueda en tiempo real en Twitter.
}
\examples{
\dontrun{
getUrlsSearchStreaming(search = "#RStats", n_urls = 200)
}

}
\references{
Puedes encontrar más información sobre el paquete TweetScrapeR en:
\url{https://github.com/agusnieto77/TweetScraperR}
}
