% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsersFullData.R
\name{getUsersFullData}
\alias{getUsersFullData}
\title{Get Users Full Data}
\usage{
getUsersFullData(
  urls_users,
  xuser = Sys.getenv("USER"),
  xpass = Sys.getenv("PASS"),
  dir = getwd()
)
}
\arguments{
\item{urls_users}{Vector de URLs de users de los cuales se desea obtener datos.}

\item{xuser}{Nombre de usuarix de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema USER.}

\item{xpass}{Contraseña de Twitter para autenticación. Por defecto es el valor de la variable de entorno del sistema PASS.}

\item{dir}{Directorio donde se guardará el archivo RDS con los datos recopilados. Por defecto es el directorio actual.}
}
\value{
Un tibble que contiene los datos de los users recuperados.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función permite recuperar y procesar datos de usuarixs de Twitter a partir de un vector de URLs
proporcionadas. Utilizando las credenciales de unx usuarix de Twitter, la función realiza la autenticación y
extrae información detallada de cada perfil de usuarix. Los datos extraídos incluyen el nombre del usuarix,
el username, la fecha de creación del perfil, el número de publicaciones, seguidorxs, seguidxs y otros metadatos.
La función maneja posibles errores durante el proceso de recolección de datos, como tiempos de espera prolongados, y
se asegura de obtener información precisa mediante múltiples intentos si es necesario. Los datos recopilados se
devuelven en forma de un tibble y se guardan en un archivo RDS para su posterior uso.
}
\examples{
\dontrun{
getUsersFullData(urls_users = "https://x.com/estacion_erre")
}

}
