% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openTwitter.R
\name{openTwitter}
\alias{openTwitter}
\title{openTwitter {TweetScraperR}}
\usage{
openTwitter()
}
\value{
Devuelve una vista de la instancia HTML de la página de inicio de sesión de Twitter.
La información de la sesión se guarda en la variable global \code{twitter} para su uso posterior.
}
\description{
\if{html}{\out{<a href="https://lifecycle.r-lib.org/articles/stages.html#experimental" target="_blank">}}\if{html}{\out{<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg" alt="[Experimental]">}}\if{html}{\out{</a>}}

Esta función permite abrir la página de inicio de sesión de Twitter en un navegador web y
guardar la instancia HTML en el entorno global. Es útil para iniciar el proceso de autenticación
antes de realizar la recolección de datos de Twitter.

Usage
openTwitter()
}
\details{
Open Twitter Login Page
}
\examples{
\dontrun{
openTwitter()
}

}
