% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmojis.R
\name{plotEmojis}
\alias{plotEmojis}
\title{Create Bar Chart of Emoticons in Tweets}
\usage{
plotEmojis(df, top_n = 10, color = "skyblue")
}
\arguments{
\item{df}{Un dataframe que contiene una columna 'emoticones' con listas de emoticones.}

\item{top_n}{Número de emoticones más frecuentes a mostrar (por defecto 10).}

\item{color}{Color de las barras en el gráfico (por defecto "skyblue").}
}
\value{
Un objeto ggplot con el gráfico de barras.
}
\description{
Esta función toma un dataframe de tweets y crea un gráfico de barras
mostrando la frecuencia de los emoticones utilizados en los tweets.
}
\examples{

df <- data.frame(emoticones = I(list(c("😊", "😂"), c("😊"), c("😂", "😍"), character(0))))
plot_emojis(df)

}
